+++ projects/app/src/app/__demo__/pages/actions-menu/actions-menu.component.ts
@@ -4,12 +4,9 @@ import {
   matTaskAlt,
   matUndo,
 } from '@ng-icons/material-icons/baseline';
+import { ButtonComponent } from '@ui/components/button';
 
-import {
-  ACTIONS_MENU_EXPORTS,
-  ActionsMenuItem,
-  ButtonComponent,
-} from '@app/common/components';
+import { ACTIONS_MENU_EXPORTS, ActionsMenuItem } from '@app/common/components';
 
 @Component({
   selector: 'app-demo-actions-menu',
