+++ projects/app/src/app/__demo__/pages/button/button.component.ts
@@ -1,8 +1,7 @@
 import { Component } from '@angular/core';
 import { NgIcon } from '@ng-icons/core';
 import { matSearch } from '@ng-icons/material-icons/baseline';
-
-import { ButtonComponent } from '@app/common/components';
+import { ButtonComponent } from '@ui/components/button';
 
 @Component({
   selector: 'app-demo-button',
