+++ projects/app/src/app/__demo__/pages/modal/modal.component.ts
@@ -1,6 +1,6 @@
 import { Component, inject } from '@angular/core';
+import { ButtonComponent } from '@ui/components/button';
 
-import { ButtonComponent } from '@app/common/components';
 import { ModalService } from '@app/common/components/modal';
 import { ModalOneComponent, ModalOneInput, ModalOneOutput } from './modal-one';
