--- projects/app/src/app/common/components/button/types.ts
@@ -1,19 +1,0 @@
-export type ButtonColor = (
-  | 'primary'
-  | 'secondary'
-  | 'tertiary'
-  | 'outline'
-  | 'ghost'
-);
-
-export type ButtonSize = (
-  | 'extra-small'
-  | 'small'
-  | 'medium'
-  | 'large'
-);
-
-export type ButtonFloatingType = (
-  | 'container'
-  | 'fixed'
-);
