+++ projects/app/src/app/common/components/index.ts
@@ -1,11 +1,9 @@
 export * from './autocomplete';
 export * from './bottom-menu';
-export * from './button';
 export * from './card-list';
 export * from './change-category-modal';
 export * from './confirm-prompt-modal';
 export * from './form-field';
-export * from './icon-button';
 export * from './linear-spinner';
 export * from './menu/actions-menu';
 export * from './modal';
