+++ projects/app/src/app/common/components/modal/modal-host/modal-host.component.ts
@@ -18,9 +18,9 @@ import {
   matCheck,
 } from '@ng-icons/material-icons/baseline';
 import { TranslocoModule } from '@jsverse/transloco';
+import { ButtonComponent } from '@ui/components/button';
+import { IconButtonComponent } from '@ui/components/icon-button';
 
-import { ButtonComponent } from '../../button';
-import { IconButtonComponent } from '../../icon-button';
 import { ModalService } from '../modal.service';
 import { ModalHostLabels } from '../types';
 import {
