+++ projects/app/src/app/common/components/quick-number/quick-number.component.ts
@@ -19,9 +19,11 @@ import {
 } from '@angular/forms';
 import { NgIcon } from '@ng-icons/core';
 import { matRemove, matAdd } from '@ng-icons/material-icons/baseline';
-
 import { uniqueId } from '@common/utils';
-import { IconButtonColor, IconButtonComponent } from '../icon-button';
+import {
+  IconButtonColor,
+  IconButtonComponent,
+} from '@ui/components/icon-button';
 
 const QUICK_NUMBER_FORM_PROVIDER: Provider = {
   provide: NG_VALUE_ACCESSOR,
