+++ projects/app/src/app/common/components/textarea/textarea.component.ts
@@ -19,12 +19,11 @@ import { NgIcon } from '@ng-icons/core';
 import { matCheck, matClear } from '@ng-icons/material-icons/baseline';
 import { FormFieldStatus } from '@common/types';
 import { cssClassesList, uniqueId } from '@common/utils';
-
+import { IconButtonComponent } from '@ui/components/icon-button';
 import {
   HTMLAttributes,
   createAttributesController,
 } from '@common/controllers';
-import { IconButtonComponent } from '../icon-button';
 
 const TEXTAREA_FORM_PROVIDER: Provider = {
   provide: NG_VALUE_ACCESSOR,
