+++ projects/app/src/app/features/inventory/components/item-form-modal/item-form-modal.component.ts
@@ -18,13 +18,13 @@ import { Observable, Subject, takeUntil } from 'rxjs';
 import { FormOption } from '@common/types';
 import { getFieldDescriptor as fDescribe } from '@common/utils';
 import { MediaQueryService } from '@ui/services';
+import { ButtonComponent } from '@ui/components/button';
 
 import {
   AUTOCOMPLETE_EXPORTS,
   AutocompleteAsyncOptionsFn,
   AutocompleteOption,
   BaseModalComponent,
-  ButtonComponent,
   FORM_FIELD_EXPORTS,
   ModalFooterDirective,
   ModalHeaderDirective,
