+++ projects/app/src/app/features/user/pages/profile/profile.component.ts
@@ -1,8 +1,9 @@
 import { DatePipe } from '@angular/common';
 import { Component, OnInit, computed, inject } from '@angular/core';
 import { TranslocoModule, TranslocoService } from '@jsverse/transloco';
+import { ButtonComponent } from '@ui/components/button';
 
-import { ButtonComponent, SelectComponent } from '@app/common/components';
+import { SelectComponent } from '@app/common/components';
 import { StackedLayoutService } from '@app/common/layouts';
 import { Language } from '@app/core/language';
 import { Theme } from '@app/core/theme';
