+++ projects/ui/src/components/button/types.ts
@@ -1,0 +1,8 @@
+export type ButtonColor =
+  | 'primary'
+  | 'secondary'
+  | 'tertiary'
+  | 'outline'
+  | 'ghost';
+
+export type ButtonSize = 'extra-small' | 'small' | 'medium' | 'large';
